Make sure stderr streaming works

+++ src/neubot/runner.go
@@ -4,14 +4,13 @@ import (
 	"errors"
 	"github.com/nu7hatch/gouuid"
 	"io/ioutil"
-	"log"
-	"os/exec"
 	"os"
+	"os/exec"
 	"time"
 )
 
 func open_file(workdir string, nettest_name string, file_type string) (
-		*os.File, error) {
+	*os.File, error) {
 	uuid4, err := uuid.NewV4()
 	if err != nil {
 		return nil, err
@@ -25,13 +24,15 @@ func open_file(workdir string, nettest_name string, file_type string) (
 }
 
 type Runner struct {
-	Status string
-	Error error
-	Process *exec.Cmd
+	Error      error
+	Process    *exec.Cmd
+	Status     string
+	StdoutPath string
+	StderrPath string
 }
 
 func RunnerStart(nettest_name string, cmdline []string, workdir string) (
-		Runner, error) {
+	Runner, error) {
 	var runner Runner
 	if len(cmdline) == 0 {
 		return runner, errors.New("invalid command line")
@@ -42,11 +43,13 @@ func RunnerStart(nettest_name string, cmdline []string, workdir string) (
 	if err != nil {
 		return runner, err
 	}
+	runner.StdoutPath = stdout.Name()
 	runner.Process.Stdout = stdout
 	stderr, err := open_file(workdir, nettest_name, "stderr")
 	if err != nil {
 		return runner, err
 	}
+	runner.StderrPath = stderr.Name()
 	runner.Process.Stderr = stderr
 	err = runner.Process.Start()
 	if err != nil {
@@ -56,18 +59,25 @@ func RunnerStart(nettest_name string, cmdline []string, workdir string) (
 	return runner, nil
 }
 
-func RunnerWait(runner Runner, timeout time.Duration) (chan error) {
+func RunnerWaitAsync(runner Runner, timeout time.Duration,
+	periodic func(Runner)) chan error {
 	done := make(chan error, 1)
 	// See: <http://stackoverflow.com/a/11886829>
 	internal := make(chan error, 1)
 	go func() {
 		internal <- runner.Process.Wait()
 	}()
+	begin := time.Now()
+again:
+	periodic(runner)
 	select {
-	case <-time.After(timeout):
-		log.Printf("kill process that is running for too long\n")
+	case <-time.After(1.0 * time.Second):
+		delta := time.Since(begin)
+		if delta < timeout {
+			goto again
+		}
 		err := runner.Process.Process.Kill()
-		if (err != nil) {
+		if err != nil {
 			// My understanding is that it should not happen, and that ot
 			// would also lead to a leak of the process control struct
 			runner.Status = "kill-failed"
@@ -76,11 +86,11 @@ func RunnerWait(runner Runner, timeout time.Duration) (chan error) {
 			err = <-internal // Make sure we don't leak control struct
 			runner.Status = "killed"
 		}
-		done <-err
+		done <- err
 	case err := <-internal:
-		log.Printf("Process terminated; exit code: %s\n", err)
-		runner.Status = "terminated"
-		done <-err
+		runner.Status = "exited"
+		done <- err
 	}
+	periodic(runner)
 	return done
 }
